+++ CodeTester-IntelliJ-IDEA-UITest/src/test/kotlin/de/fxnm/fixtures/IdeaFrame.kt
@@ -64,9 +64,8 @@ class IdeaFrame(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) :
 
 
     fun showCodeTesterExplorer() {
-        try {
-            find<CodeTesterExplorer>(byXpath("//div[@class='CodeTesterToolWindowPanel']"))
-        } catch (e: Exception) {
+        // FIXME: 18.01.2021 BUG IF CURRENT OPEN TOOL WINDOW IS NOT THE MAIN TOOL WINDOW
+        if (findAll<CodeTesterExplorer>(CodeTesterExplorer::class.java).isEmpty()) {
             find(ComponentFixture::class.java,
                  byXpath("//div[@accessiblename='CodeTester' and @class='StripeButton' and @text='CodeTester']")).click()
         }