+++ CodeTester-IntelliJ-IDEA-UITest/src/test/kotlin/de/fxnm/tests/RunCodeTester.kt
@@ -2,6 +2,8 @@ package de.fxnm.tests
 
 import com.intellij.remoterobot.stepsProcessing.step
 import de.fxnm.extensions.uiTest
+import de.fxnm.fixtures.checkResultTree
+import de.fxnm.fixtures.codeTesterTab
 import de.fxnm.fixtures.codeTesterToolWindow
 import de.fxnm.fixtures.idea
 import de.fxnm.steps.CodeTesterToolWindowSteps
@@ -18,14 +20,24 @@ class RunCodeTester : UiTestcaseBase() {
         uiTest {
             idea {
                 showCodeTesterExplorer()
+
                 codeTesterToolWindow {
                     step("Run Test") {
-//                        selectCategory("2020 SS Final 2")
+                        selectCategory("2020 SS Final 2")
                         run()
 
-                        // TODO: 26.12.2020
                         waitForBackgroundTasks()
+
+                        checkResultTree {
+                            doubleClickItem("Check 'Feiertage-Test (Fehlerhaftes Format der Datei)' : FAILED")
+                        }
                     }
+                }
+
+                codeTesterTab {
+                    selectMainPanel()
+                    selectPanel("Check 'Feiertage-Test (Fehlerhaftes Format der Datei)' : FAILED")
+                    selectMainPanel()
                 }
             }
         }